Fix x64 build error

+++ PowerEditor/src/WinControls/PluginsAdmin/pluginsAdmin.cpp
@@ -610,7 +610,7 @@ void PluginViewList::pushBack(PluginUpdateInfo* pi)
 
 	// add in order
 	size_t i = _ui.findAlphabeticalOrderPos(pi->_displayName, _sortType == DISPLAY_NAME_ALPHABET_ENCREASE ? _ui.sortEncrease : _ui.sortDecrease);
-	_ui.addLine(values2Add, reinterpret_cast<LPARAM>(pi), i);
+	_ui.addLine(values2Add, reinterpret_cast<LPARAM>(pi), static_cast<int>(i));
 }
 
 bool loadFromJson(PluginViewList & pl, const json& j)
